+++ src/nplinker/genomics/antismash/__init__.py
@@ -2,8 +2,18 @@ import logging
 from .antismash_downloader import download_and_extract_antismash_data
 from .antismash_loader import AntismashBGCLoader
 from .antismash_loader import parse_bgc_genbank
+from .podp_antismash_downloader import GenomeStatus
+from .podp_antismash_downloader import get_best_available_genome_id
+from .podp_antismash_downloader import podp_download_and_extract_antismash_data
 
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
-__all__ = ["AntismashBGCLoader", "parse_bgc_genbank", "download_and_extract_antismash_data"]
+__all__ = [
+    "download_and_extract_antismash_data",
+    "AntismashBGCLoader",
+    "parse_bgc_genbank",
+    "GenomeStatus",
+    "get_best_available_genome_id",
+    "podp_download_and_extract_antismash_data",
+]
