+++ src/nplinker/pairedomics/__init__.py
@@ -1,7 +1,4 @@
 import logging
-from .podp_antismash_downloader import GenomeStatus
-from .podp_antismash_downloader import get_best_available_genome_id
-from .podp_antismash_downloader import podp_download_and_extract_antismash_data
 from .strain_mappings_generator import extract_mappings_ms_filename_spectrum_id
 from .strain_mappings_generator import extract_mappings_original_genome_id_resolved_genome_id
 from .strain_mappings_generator import extract_mappings_resolved_genome_id_bgc_id
@@ -15,9 +12,6 @@ from .strain_mappings_generator import podp_generate_strain_mappings
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
 __all__ = [
-    "GenomeStatus",
-    "get_best_available_genome_id",
-    "podp_download_and_extract_antismash_data",
     "podp_generate_strain_mappings",
     "extract_mappings_strain_id_original_genome_id",
     "extract_mappings_original_genome_id_resolved_genome_id",
