+++ src/nplinker/pairedomics/strain_mappings_generator.py
@@ -8,8 +8,8 @@ from nplinker.schemas import GENOME_BGC_MAPPINGS_SCHEMA
 from nplinker.schemas import validate_podp_json
 from nplinker.strain import Strain
 from nplinker.strain_collection import StrainCollection
-from .podp_antismash_downloader import GenomeStatus
-from .podp_antismash_downloader import get_best_available_genome_id
+from ..genomics.antismash.podp_antismash_downloader import GenomeStatus
+from ..genomics.antismash.podp_antismash_downloader import get_best_available_genome_id
 
 
 logger = logging.getLogger(__name__)
